SeedIOT
The application has access to instant and historical data of the sensors via the thingspeak platform and shows them to the user in the application. These consist of a total of 4 data, namely node 1 temperature and humidity, node 2 temperature and humidity. In addition, the past 10-day data is received, and the averages are shown to the user. Past 10-day Node 1 temperature data are graphically displayed to the user.